requirements.txt
streamlit>=1.24.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.13.0
scikit-learn>=1.0.0
transformers>=4.30.0
laser-encoders>=0.0.4
laserembeddings[zh]>=1.1.1
umap-learn>=0.5.3
phate>=1.0.7
python-dotenv>=0.19.0
requests>=2.31.0
sentencepiece>=0.1.99